html/semantics/embedded-content-0/media-elements/error_null.html

<!doctype html>
<html>
 <head>
  <title>{audio,video}.error - null if no source</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#dom-media-error">spec reference</a></p>
  <audio id="a">
  </audio>
  <video id="v">
  </video>
  <div id="log"></div>
  <script>
test(function() {
 assert_equals(
  document.getElementById("a").error,
  null,
  "audioElement.error property is null if no source");
}, "audio.error - null if no source");
test(function() {
 assert_equals(
  document.getElementById("v").error,
  null,
  "videoElement.error property is null if no source");
}, "video.error - null if no source");
  </script>
 </body>
</html>
